Revert "[fix][io] Upgrade debezium oracle to 2.5.4.Final  (#279)"
This reverts commit 56424a1ec7108597ae76912efa507f7ccf401a78.
