+++ pom.xml
@@ -164,8 +164,7 @@ flexible messaging model and an intuitive client API.</description>
     <presto.version>334</presto.version>
     <scala.binary.version>2.13</scala.binary.version>
     <scala-library.version>2.13.10</scala-library.version>
-    <debezium.version>1.9.7.Final</debezium.version>
-    <debezium.oracle.version>2.5.4.Final</debezium.oracle.version>
+    <debezium.version>2.5.4.Final</debezium.version>
     <debezium.postgresql.version>42.5.0</debezium.postgresql.version>
     <debezium.mysql.version>8.0.30</debezium.mysql.version>
     <!-- Override version that brings CVE-2022-3143 with debezium -->
